-Set up dev repo and linked to it on the front page. -Converted page to Python / SUIT for any server-sid work I might need. -For an example of why I might need it, see the new "Latest Changeset" section. It displays the latest Mercurial changeset, effectively providing viewers with the latest updates on the page without any hassle on my end.
Later:
-Must fix mobile.js bug. It appears that clicking on either of the repo links causes some kind of hook error.
-Will continue trying to eliminate impasse cases with Random and Shuffle.

I am liking Mecurial so far. This setup allows me to develop Hanoi and keep both a stable and development copy of the page, so now, without fear of taking down the page for repairs, I will finally link the page on my site. I kind of wish I used this from the beginning, but most of the revisions up to this point have been direct conversions of Fred Lunnon's original program, so all that was lost in the process is expendable. Definitely try Mercurial out over SVN, as from what I've seen, it beats it on every imaginable level.